--HG--
branch : spec_SFEredis-32

+++ specs/SFEperl-cpan-meta-yaml.spec
@@ -1,20 +1,22 @@
 %include Solaris.inc
-%include default-depend.inc
 
 %define build584 0
+%define build510 %( if [ -x /usr/perl5/5.10/bin/perl ]; then echo '1'; else echo '0'; fi)
 %define build512 %( if [ -x /usr/perl5/5.12/bin/perl ]; then echo '1'; else echo '0'; fi)
 %define build516 %( if [ -x /usr/perl5/5.16/bin/perl ]; then echo '1'; else echo '0'; fi)
-%define build520 %( if [ -x /usr/perl5/5.20/bin/perl ]; then echo '1'; else echo '0'; fi)
+%define build522 %( if [ -x /usr/perl5/5.22/bin/perl ]; then echo '1'; else echo '0'; fi)
+%define enable_test %( if [ "x${PERL_DISABLE_TEST}" = 'xtrue' ]; then echo '0'; else echo '1'; fi )
 %define include_executable 0
 
 %define cpan_name CPAN-Meta-YAML
 %define sfe_cpan_name cpan-meta-yaml
+%define ips_cpan_name cpan-meta-yaml
 
 Summary:               Read and write a subset of YAML for CPAN Meta files
 Name:                  SFEperl-%{sfe_cpan_name}
-IPS_package_name:      library/perl-5/%{sfe_cpan_name}
-Version:               0.016
-IPS_component_version: 0.16
+IPS_package_name:      library/perl-5/%{ips_cpan_name}
+Version:               0.018
+IPS_component_version: 0.18
 License:               perl_5
 URL:                   https://metacpan.org/pod/CPAN::Meta::YAML
 Source0:               http://cpan.metacpan.org/authors/id/D/DA/DAGOLDEN/CPAN-Meta-YAML-%{version}.tar.gz
@@ -25,42 +27,77 @@ Read and write a subset of YAML for CPAN Meta files
 
 %if %{build584}
 %package 584
-IPS_package_name: library/perl-5/%{sfe_cpan_name}-584
+IPS_package_name: library/perl-5/%{ips_cpan_name}-584
 Summary:          Read and write a subset of YAML for CPAN Meta files
 BuildRequires:    runtime/perl-584 = *
 BuildRequires:    library/perl-5/cpan-meta-584
 BuildRequires:    library/perl-5/extutils-makemaker-584
+BuildRequires:    library/perl-5/file-temp-584
 BuildRequires:    library/perl-5/io-584
 BuildRequires:    library/perl-5/json-pp-584
 BuildRequires:    library/perl-5/pathtools-584
 BuildRequires:    library/perl-5/test-simple-584
+%if %{enable_test}
+BuildRequires:    library/perl-5/carp-584
+BuildRequires:    library/perl-5/exporter-584
+BuildRequires:    library/perl-5/scalar-list-utils-584
+%endif
 Requires:         runtime/perl-584 = *
-Requires:         library/perl-5/b-584
+Requires:         library/perl-5/%{ips_cpan_name}
 Requires:         library/perl-5/carp-584
 Requires:         library/perl-5/exporter-584
-Requires:         library/perl-5/fcntl-584
 Requires:         library/perl-5/scalar-list-utils-584
 
 %description 584
 Read and write a subset of YAML for CPAN Meta files
 %endif
 
+%if %{build510}
+%package 510
+IPS_package_name: library/perl-5/%{ips_cpan_name}-510
+Summary:          Read and write a subset of YAML for CPAN Meta files
+BuildRequires:    runtime/perl-510 = *
+BuildRequires:    library/perl-5/cpan-meta-510
+BuildRequires:    library/perl-5/extutils-makemaker-510
+BuildRequires:    library/perl-5/file-temp-510
+BuildRequires:    library/perl-5/io-510
+BuildRequires:    library/perl-5/json-pp-510
+BuildRequires:    library/perl-5/pathtools-510
+BuildRequires:    library/perl-5/test-simple-510
+BuildRequires:    library/perl-5/carp-510
+BuildRequires:    library/perl-5/exporter-510
+BuildRequires:    library/perl-5/scalar-list-utils-510
+Requires:         runtime/perl-510 = *
+Requires:         library/perl-5/%{ips_cpan_name}
+Requires:         library/perl-5/carp-510
+Requires:         library/perl-5/exporter-510
+Requires:         library/perl-5/scalar-list-utils-510
+
+%description 510
+Read and write a subset of YAML for CPAN Meta files
+%endif
+
 %if %{build512}
 %package 512
-IPS_package_name: library/perl-5/%{sfe_cpan_name}-512
+IPS_package_name: library/perl-5/%{ips_cpan_name}-512
 Summary:          Read and write a subset of YAML for CPAN Meta files
 BuildRequires:    runtime/perl-512 = *
 BuildRequires:    library/perl-5/cpan-meta-512
 BuildRequires:    library/perl-5/extutils-makemaker-512
+BuildRequires:    library/perl-5/file-temp-512
 BuildRequires:    library/perl-5/io-512
 BuildRequires:    library/perl-5/json-pp-512
 BuildRequires:    library/perl-5/pathtools-512
 BuildRequires:    library/perl-5/test-simple-512
+%if %{enable_test}
+BuildRequires:    library/perl-5/carp-512
+BuildRequires:    library/perl-5/exporter-512
+BuildRequires:    library/perl-5/scalar-list-utils-512
+%endif
 Requires:         runtime/perl-512 = *
-Requires:         library/perl-5/b-512
+Requires:         library/perl-5/%{ips_cpan_name}
 Requires:         library/perl-5/carp-512
 Requires:         library/perl-5/exporter-512
-Requires:         library/perl-5/fcntl-512
 Requires:         library/perl-5/scalar-list-utils-512
 
 %description 512
@@ -69,51 +106,63 @@ Read and write a subset of YAML for CPAN Meta files
 
 %if %{build516}
 %package 516
-IPS_package_name: library/perl-5/%{sfe_cpan_name}-516
+IPS_package_name: library/perl-5/%{ips_cpan_name}-516
 Summary:          Read and write a subset of YAML for CPAN Meta files
 BuildRequires:    runtime/perl-516 = *
 BuildRequires:    library/perl-5/cpan-meta-516
 BuildRequires:    library/perl-5/extutils-makemaker-516
+BuildRequires:    library/perl-5/file-temp-516
 BuildRequires:    library/perl-5/io-516
 BuildRequires:    library/perl-5/json-pp-516
 BuildRequires:    library/perl-5/pathtools-516
 BuildRequires:    library/perl-5/test-simple-516
+Requires:         library/perl-5/%{ips_cpan_name}
+%if %{enable_test}
+BuildRequires:    library/perl-5/carp-516
+BuildRequires:    library/perl-5/exporter-516
+BuildRequires:    library/perl-5/scalar-list-utils-516
+%endif
 Requires:         runtime/perl-516 = *
-Requires:         library/perl-5/b-516
+Requires:         library/perl-5/%{ips_cpan_name}
 Requires:         library/perl-5/carp-516
 Requires:         library/perl-5/exporter-516
-Requires:         library/perl-5/fcntl-516
 Requires:         library/perl-5/scalar-list-utils-516
 
 %description 516
 Read and write a subset of YAML for CPAN Meta files
 %endif
 
-%if %{build520}
-%package 520
-IPS_package_name: library/perl-5/%{sfe_cpan_name}-520
+%if %{build522}
+%package 522
+IPS_package_name: library/perl-5/%{ips_cpan_name}-522
 Summary:          Read and write a subset of YAML for CPAN Meta files
-BuildRequires:    runtime/perl-520 = *
-BuildRequires:    library/perl-5/cpan-meta-520
-BuildRequires:    library/perl-5/extutils-makemaker-520
-BuildRequires:    library/perl-5/io-520
-BuildRequires:    library/perl-5/json-pp-520
-BuildRequires:    library/perl-5/pathtools-520
-BuildRequires:    library/perl-5/test-simple-520
-Requires:         runtime/perl-520 = *
-Requires:         library/perl-5/b-520
-Requires:         library/perl-5/carp-520
-Requires:         library/perl-5/exporter-520
-Requires:         library/perl-5/fcntl-520
-Requires:         library/perl-5/scalar-list-utils-520
+BuildRequires:    runtime/perl-522 = *
+BuildRequires:    library/perl-5/cpan-meta-522
+BuildRequires:    library/perl-5/extutils-makemaker-522
+BuildRequires:    library/perl-5/file-temp-522
+BuildRequires:    library/perl-5/io-522
+BuildRequires:    library/perl-5/json-pp-522
+BuildRequires:    library/perl-5/pathtools-522
+BuildRequires:    library/perl-5/test-simple-522
+%if %{enable_test}
+BuildRequires:    library/perl-5/carp-522
+BuildRequires:    library/perl-5/exporter-522
+BuildRequires:    library/perl-5/scalar-list-utils-522
+%endif
+Requires:         runtime/perl-522 = *
+Requires:         library/perl-5/%{ips_cpan_name}
+Requires:         library/perl-5/carp-522
+Requires:         library/perl-5/exporter-522
+Requires:         library/perl-5/scalar-list-utils-522
 
-%description 520
+%description 522
 Read and write a subset of YAML for CPAN Meta files
 %endif
 
+
 %prep
 %setup -q -n %{cpan_name}-%{version}
-rm -rf %{buildroot}
+[ -d %{buildroot} ] && rm -rf %{buildroot}
 
 %build
 build_with_makefile.pl_for() {
@@ -126,8 +175,12 @@ build_with_makefile.pl_for() {
     ${bindir}/perl Makefile.PL PREFIX=%{_prefix} \
                    DESTDIR=$RPM_BUILD_ROOT \
                    LIB=${vendor_dir}
-    make
-    [ ${test} == 'without_test' ] || make test
+
+    export CC='cc -m32'
+    export LD='cc -m32'
+    echo ${perl_ver} | egrep '5\.(84|12)' > /dev/null || (export CC='cc -m64'; export LD='cc -m64')
+    make CC="${CC}" LD="${LD}"
+    [ "x${PERL_DISABLE_TEST}" = 'xtrue' ] || [ "x${test}" = 'xwithout_test' ] || make test CC="${CC}" "LD=${LD}"
     make pure_install
 }
 
@@ -142,35 +195,70 @@ build_with_build.pl_for() {
                    --installdirs vendor \
                    --destdir $RPM_BUILD_ROOT
     ${bindir}/perl ./Build
-    [ ${test} == 'without_test' ] || ${bindir}/perl ./Build test
+    [ "x${PERL_DISABLE_TEST}" = 'xtrue' ] || [ "x${test}" = 'xwithout_test' ] || ${bindir}/perl ./Build test
     ${bindir}/perl ./Build install --destdir $RPM_BUILD_ROOT
+    ${bindir}/perl ./Build clean
 }
 
 modify_bin_dir() {
-  perl_ver=$1
-  if [ -d $RPM_BUILD_ROOT/usr/bin ]
-  then
-    [ -d $RPM_BUILD_ROOT/usr/perl5/${perl_ver} ] || mkdir -p $RPM_BUILD_ROOT/usr/perl5/${perl_ver}
-    mv $RPM_BUILD_ROOT/usr/bin $RPM_BUILD_ROOT/usr/perl5/${perl_ver}/bin
-  fi
+    perl_ver=$1
+    if [ -d $RPM_BUILD_ROOT/usr/bin ]
+    then
+      [ -d $RPM_BUILD_ROOT/usr/perl5/${perl_ver} ] || mkdir -p $RPM_BUILD_ROOT/usr/perl5/${perl_ver}
+      mv $RPM_BUILD_ROOT/usr/bin $RPM_BUILD_ROOT/usr/perl5/${perl_ver}/bin
+    fi
+      
+    if [ -d $RPM_BUILD_ROOT/usr/perl5/${perl_ver}/bin ]
+    then
+        for i in $RPM_BUILD_ROOT/usr/perl5/${perl_ver}/bin/*
+        do
+            sed -i.bak -e "s!/usr/bin/env perl!/usr/perl5/${perl-ver}/bin/perl!" ${i}
+            [ -f ${i}.bak] || rm -f ${i}.bak
+        done
+    fi
+}
+
+modify_man_dir() {
+    perl_ver=$1
+    if [ -d $RPM_BUILD_ROOT/usr/perl5/${perl_ver}/man ]
+    then
+        if [ -d $RPM_BUILD_ROOT%{_datadir}/man ]
+        then
+            rm -rf $RPM_BUILD_ROOT/usr/perl5/${perl_ver}/man
+        else
+            mkdir -p $RPM_BUILD_ROOT%{_datadir}
+            mv $RPM_BUILD_ROOT/usr/perl5/${perl_ver}/man $RPM_BUILD_ROOT%{_datadir}/
+            rm -rf $RPM_BUILD_ROOT/usr/perl5/${perl_ver}/man
+        fi
+        if [ %{include_executable} -eq 0 ]
+        then
+            rmdir $RPM_BUILD_ROOT/usr/perl5/${perl_ver}
+        fi
+
+    fi
 }
 
 build_for() {
-  if [ -f Makefile.PL ];
-  then
-    build_with_makefile.pl_for $*
-  elif [ -f Build.PL ];
+  if [ -f Build.PL ];
   then
     build_with_build.pl_for $*
+  elif [ -f Makefile.PL ];
+  then
+    build_with_makefile.pl_for $*
   fi
 
-    modify_bin_dir $*
+  modify_bin_dir $*
+  modify_man_dir $*
 }
 
 # To build without test, pass 'without_test' to build_for commaond.
 # like 'build_for version without_test'
 %if %{build584}
 build_for 5.8.4
+%endif
+
+%if %{build510}
+build_for 5.10
 %endif
 
 %if %{build512}
@@ -181,14 +269,14 @@ build_for 5.12
 build_for 5.16
 %endif
 
-%if %{build520}
-build_for 5.20
+%if %{build522}
+build_for 5.22
 %endif
 
 %install
-mkdir -p $RPM_BUILD_ROOT%{_datadir}
 if [ -d $RPM_BUILD_ROOT%{_prefix}/man ]
 then
+    mkdir -p $RPM_BUILD_ROOT%{_datadir}
     mv $RPM_BUILD_ROOT%{_prefix}/man $RPM_BUILD_ROOT%{_datadir}
 fi
 if [ -d $RPM_BUILD_ROOT%{_datadir}/man/man3 ]
@@ -201,7 +289,7 @@ rm -rf %{buildroot}
 
 %files
 %defattr(0755,root,bin,-)
-%{_datadir}/man/man3perl
+%{_datadir}/man
 
 %if %{build584}
 %files 584
@@ -210,6 +298,16 @@ rm -rf %{buildroot}
 /usr/perl5/vendor_perl/5.8.4
 %if %{include_executable}
 /usr/perl5/5.8.4
+%endif
+%endif
+
+%if %{build510}
+%files 510
+%defattr(0755,root,bin,-)
+%dir %attr (0755, root, sys) /usr
+/usr/perl5/vendor_perl/5.10
+%if %{include_executable}
+/usr/perl5/5.1.0
 %endif
 %endif
 
@@ -233,24 +331,23 @@ rm -rf %{buildroot}
 %endif
 %endif
 
-%if %{build520}
-%files 520
+%if %{build522}
+%files 522
 %defattr(0755,root,bin,-)
 %dir %attr (0755, root, sys) /usr
-/usr/perl5/vendor_perl/5.20
+/usr/perl5/vendor_perl/5.22
 %if %{include_executable}
-/usr/perl5/5.20
+/usr/perl5/5.22
 %endif
 %endif
-
 
 %changelog
+* Thu Jun 01 2017 - Fumihisa TONAKA <fumi.ftnk@gmail.com>
+- bump to 0.018
 * Sun Nov 08 2015 - Fumihisa TONAKA <fumi.ftnk@gmail.com>
 - update Requires
 * Wed Nov 04 2015 - Fumihisa TONAKA <fumi.ftnk@gmail.com>
 - bump to 0.016 and build packages for perl-516 and perl-520
-* Sun Dec 07 2014 - Fumihisa TONAKA <fumi.ftnk@gmail.com>
-- add Buildrequires
 * Wed Nov 19 2014 - Fumihisa TONAKA <fumi.ftnk@gmail.com>
 - add BuildRequires
 - bump to 0.012
