+++ specs/SFEperl-encode.spec
@@ -4,8 +4,9 @@
 %define build510 %( if [ -x /usr/perl5/5.10/bin/perl ]; then echo '1'; else echo '0'; fi)
 %define build512 %( if [ -x /usr/perl5/5.12/bin/perl ]; then echo '1'; else echo '0'; fi)
 %define build516 %( if [ -x /usr/perl5/5.16/bin/perl ]; then echo '1'; else echo '0'; fi)
-%define build520 %( if [ -x /usr/perl5/5.20/bin/perl ]; then echo '1'; else echo '0'; fi)
-%define include_executable 1
+%define build522 %( if [ -x /usr/perl5/5.22/bin/perl ]; then echo '1'; else echo '0'; fi)
+%define enable_test %( if [ "x${PERL_DISABLE_TEST}" = 'xtrue' ]; then echo '0'; else echo '1'; fi )
+%define include_executable 0
 
 %define cpan_name Encode
 %define sfe_cpan_name encode
@@ -14,8 +15,8 @@
 Summary:               character encodings in Perl
 Name:                  SFEperl-%{sfe_cpan_name}
 IPS_package_name:      library/perl-5/%{ips_cpan_name}
-Version:               2.78
-IPS_component_version: 2.78
+Version:               2.89
+IPS_component_version: 2.89
 License:               perl_5
 URL:                   https://metacpan.org/pod/Encode
 Source0:               http://cpan.metacpan.org/authors/id/D/DA/DANKOGAI/Encode-%{version}.tar.gz
@@ -30,6 +31,11 @@ IPS_package_name: library/perl-5/%{ips_cpan_name}-584
 Summary:          character encodings in Perl
 BuildRequires:    runtime/perl-584 = *
 BuildRequires:    library/perl-5/extutils-makemaker-584
+BuildRequires:    library/perl-5/test-simple-584
+%if %{enable_test}
+BuildRequires:    library/perl-5/exporter-584
+BuildRequires:    library/perl-5/parent-584
+%endif
 Requires:         runtime/perl-584 = *
 Requires:         library/perl-5/%{ips_cpan_name}
 Requires:         library/perl-5/exporter-584
@@ -45,6 +51,9 @@ IPS_package_name: library/perl-5/%{ips_cpan_name}-510
 Summary:          character encodings in Perl
 BuildRequires:    runtime/perl-510 = *
 BuildRequires:    library/perl-5/extutils-makemaker-510
+BuildRequires:    library/perl-5/test-simple-510
+BuildRequires:    library/perl-5/exporter-510
+BuildRequires:    library/perl-5/parent-510
 Requires:         runtime/perl-510 = *
 Requires:         library/perl-5/%{ips_cpan_name}
 Requires:         library/perl-5/exporter-510
@@ -60,6 +69,11 @@ IPS_package_name: library/perl-5/%{ips_cpan_name}-512
 Summary:          character encodings in Perl
 BuildRequires:    runtime/perl-512 = *
 BuildRequires:    library/perl-5/extutils-makemaker-512
+BuildRequires:    library/perl-5/test-simple-512
+%if %{enable_test}
+BuildRequires:    library/perl-5/exporter-512
+BuildRequires:    library/perl-5/parent-512
+%endif
 Requires:         runtime/perl-512 = *
 Requires:         library/perl-5/%{ips_cpan_name}
 Requires:         library/perl-5/exporter-512
@@ -75,6 +89,12 @@ IPS_package_name: library/perl-5/%{ips_cpan_name}-516
 Summary:          character encodings in Perl
 BuildRequires:    runtime/perl-516 = *
 BuildRequires:    library/perl-5/extutils-makemaker-516
+BuildRequires:    library/perl-5/test-simple-516
+Requires:         library/perl-5/%{ips_cpan_name}
+%if %{enable_test}
+BuildRequires:    library/perl-5/exporter-516
+BuildRequires:    library/perl-5/parent-516
+%endif
 Requires:         runtime/perl-516 = *
 Requires:         library/perl-5/%{ips_cpan_name}
 Requires:         library/perl-5/exporter-516
@@ -84,25 +104,30 @@ Requires:         library/perl-5/parent-516
 character encodings in Perl
 %endif
 
-%if %{build520}
-%package 520
-IPS_package_name: library/perl-5/%{ips_cpan_name}-520
+%if %{build522}
+%package 522
+IPS_package_name: library/perl-5/%{ips_cpan_name}-522
 Summary:          character encodings in Perl
-BuildRequires:    runtime/perl-520 = *
-BuildRequires:    library/perl-5/extutils-makemaker-520
-Requires:         runtime/perl-520 = *
+BuildRequires:    runtime/perl-522 = *
+BuildRequires:    library/perl-5/extutils-makemaker-522
+BuildRequires:    library/perl-5/test-simple-522
+%if %{enable_test}
+BuildRequires:    library/perl-5/exporter-522
+BuildRequires:    library/perl-5/parent-522
+%endif
+Requires:         runtime/perl-522 = *
 Requires:         library/perl-5/%{ips_cpan_name}
-Requires:         library/perl-5/exporter-520
-Requires:         library/perl-5/parent-520
+Requires:         library/perl-5/exporter-522
+Requires:         library/perl-5/parent-522
 
-%description 520
+%description 522
 character encodings in Perl
 %endif
 
 
 %prep
 %setup -q -n %{cpan_name}-%{version}
-rm -rf %{buildroot}
+[ -d %{buildroot} ] && rm -rf %{buildroot}
 
 %build
 build_with_makefile.pl_for() {
@@ -115,8 +140,12 @@ build_with_makefile.pl_for() {
     ${bindir}/perl Makefile.PL PREFIX=%{_prefix} \
                    DESTDIR=$RPM_BUILD_ROOT \
                    LIB=${vendor_dir}
-    make
-    [ x${test} = 'xwithout_test' ] || make test
+
+    export CC='cc -m32'
+    export LD='cc -m32'
+    echo ${perl_ver} | egrep '5\.(84|12)' > /dev/null || (export CC='cc -m64'; export LD='cc -m64')
+    make CC="${CC}" LD="${LD}"
+    [ "x${PERL_DISABLE_TEST}" = 'xtrue' ] || [ "x${test}" = 'xwithout_test' ] || make test CC="${CC}" "LD=${LD}"
     make pure_install
 }
 
@@ -131,7 +160,7 @@ build_with_build.pl_for() {
                    --installdirs vendor \
                    --destdir $RPM_BUILD_ROOT
     ${bindir}/perl ./Build
-    [ x${test} = 'xwithout_test' ] || ${bindir}/perl ./Build test
+    [ "x${PERL_DISABLE_TEST}" = 'xtrue' ] || [ "x${test}" = 'xwithout_test' ] || ${bindir}/perl ./Build test
     ${bindir}/perl ./Build install --destdir $RPM_BUILD_ROOT
     ${bindir}/perl ./Build clean
 }
@@ -146,11 +175,9 @@ modify_bin_dir() {
       
     if [ -d $RPM_BUILD_ROOT/usr/perl5/${perl_ver}/bin ]
     then
-	# enc2xs and piconv are provided by perl-512, perl-516 and perl-520
-	rm -f $RPM_BUILD_ROOT/usr/perl5/${perl_ver}/bin/{enc2xs,piconv}
         for i in $RPM_BUILD_ROOT/usr/perl5/${perl_ver}/bin/*
         do
-            sed -i.bak -e "s/\/usr\/bin\/env ruby/\/usr\/perl5\/${perl-ver}\/bin\/ruby/" ${i}
+            sed -i.bak -e "s!/usr/bin/env perl!/usr/perl5/${perl-ver}/bin/perl!" ${i}
             [ -f ${i}.bak] || rm -f ${i}.bak
         done
     fi
@@ -168,7 +195,11 @@ modify_man_dir() {
             mv $RPM_BUILD_ROOT/usr/perl5/${perl_ver}/man $RPM_BUILD_ROOT%{_datadir}/
             rm -rf $RPM_BUILD_ROOT/usr/perl5/${perl_ver}/man
         fi
-        rmdir $RPM_BUILD_ROOT/usr/perl5/${perl_ver}
+        if [ %{include_executable} -eq 0 ]
+        then
+            rmdir $RPM_BUILD_ROOT/usr/perl5/${perl_ver}
+        fi
+
     fi
 }
 
@@ -203,8 +234,8 @@ build_for 5.12
 build_for 5.16
 %endif
 
-%if %{build520}
-build_for 5.20
+%if %{build522}
+build_for 5.22
 %endif
 
 %install
@@ -217,6 +248,8 @@ if [ -d $RPM_BUILD_ROOT%{_datadir}/man/man3 ]
 then
     mv $RPM_BUILD_ROOT%{_datadir}/man/man3 $RPM_BUILD_ROOT%{_datadir}/man/man3perl
 fi
+
+rm -rf ${RPM_BUILD_ROOT}/usr/perl5/5.*
 
 %clean
 rm -rf %{buildroot}
@@ -265,18 +298,20 @@ rm -rf %{buildroot}
 %endif
 %endif
 
-%if %{build520}
-%files 520
+%if %{build522}
+%files 522
 %defattr(0755,root,bin,-)
 %dir %attr (0755, root, sys) /usr
-/usr/perl5/vendor_perl/5.20
+/usr/perl5/vendor_perl/5.22
 %if %{include_executable}
-/usr/perl5/5.20
+/usr/perl5/5.22
 %endif
 %endif
-
 
 %changelog
+* Wed Apr 26 2017 - Fumihisa TONAKA <fumi.ftnk@gmail.com>
+- bump to 2.89, package for perl-522 is added and for perl-520 is obsolete
+- not include executable files because of conflicting with files included in perl-5XX
 * Fri Nov 13 2015 - Fumihisa TONAKA <fumi.ftnk@gmail.com>
 - bump to 2.78 and build packages for perl-510, perl-516 and perl-520
 * Thu Oct 30 2013 - Fumihisa TONAKA <fumi.ftnk@gmail.com>
